requirements.txt
PyMuPDF>=1.23.0
pdfminer.six>=20221105
pdfplumber>=0.10.0
python-docx>=0.8.11
pyyaml>=6.0
regex>=2023.0.0
Unidecode>=1.3.0
